+++ src/components/pages/CreateBook/formData.js
@@ -1,0 +1,23 @@
+
+export default [
+  {
+    name: 'title',
+    type: 'text',
+    placeholder: 'Enter book title'
+  },
+  {
+    name: 'author',
+    type: 'text',
+    placeholder: 'Enter name of author'
+  },
+  {
+    name: 'publisher',
+    type: 'text',
+    placeholder: 'Enter name of publisher'
+  }, 
+  {
+    name: 'image',
+    type: 'url',
+    placeholder: 'Upload image'
+  }
+];
